small modify

+++ proc/src/main/java/com/gars/percents/home/State.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
  * Created by gars on 29.12.2016.
  */
 data class State (
-        var procents: Int = 10,
+        var procents: Float = 10f,
         var deposit: Float = 0f,
         var mounthAdd: Int = 0,
         var mounthAddBreak: Int = 0,
